job-search-automation/backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
scikit-learn>=1.4.0
pandas>=2.1.0
numpy>=1.26.0
python-docx>=1.1.0
PyPDF2>=3.0.1
beautifulsoup4>=4.12.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
requests>=2.31.0

